feat: return scores in decimal

+++ src/SchrodingerServer.Application.Contracts/AddressRelationship/Dto/RankDto.cs
@@ -13,7 +13,7 @@ public class RankDto
 public class ActivityRankData
 {
     public string Address { get; set; }
-    public long Scores { get; set; }
+    public decimal Scores { get; set; }
     public string Reward { get; set; }
     
     public DateTime UpdateTime { get; set; }
